requirements.txt
Flask
Flask-SocketIO
Flask-SQLAlchemy
python-dotenv
eventlet
shortuuid
flask-cors
psycopg2-binary
pyjwt